Move AttentionLSTM Layer into networks folder
